+++ backend/Mizan.Domain/Entities/Food.cs
@@ -8,7 +8,7 @@ public class Food
     public string? Barcode { get; set; }
     public decimal ServingSize { get; set; } = 100;
     public string ServingUnit { get; set; } = "g";
-    public int CaloriesPer100g { get; set; }
+    public decimal CaloriesPer100g { get; set; }
     public decimal ProteinPer100g { get; set; }
     public decimal CarbsPer100g { get; set; }
     public decimal FatPer100g { get; set; }
